% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_test}
\alias{mean_test}
\title{Get the mean of two values}
\usage{
mean_test(x, y)
}
\arguments{
\item{x}{a value}

\item{y}{a value}
}
\description{
This is a test function. Here more text.
}
\examples{
mean_test(1, 2)
}
